requirements.txt
pyside6
pyserial
numpy
pyqtgraph
scipy
psutil
